goals tiered reward handling
