+++ packages/client/abi/GoalClaimSystem.json
@@ -208,11 +208,6 @@
     },
     {
       "type": "error",
-      "name": "QueryOrderNotSupported",
-      "inputs": []
-    },
-    {
-      "type": "error",
       "name": "Unauthorized",
       "inputs": []
     }
