+++ packages/client/src/app/components/canvas/Scene.tsx
@@ -6,7 +6,7 @@ import { registerUIComponent } from 'app/root';
 import { useAccount, useNetwork, useSelected } from 'app/stores';
 import { backgrounds } from 'assets/images/backgrounds';
 import { getAccountByOwner } from 'network/shapes/Account';
-import { getGoalByIndex } from 'network/shapes/Goal';
+import { getGoalByIndex } from 'network/shapes/Goals';
 import { Room } from './Room';
 
 // The Scene paints the wallpaper and the room. It updates the selected room
