+++ packages/client/src/app/components/modals/goals/Progress.tsx
@@ -4,7 +4,7 @@ import styled from 'styled-components';
 import { EntityIndex } from '@mud-classic/recs';
 import { ProgressBar } from 'app/components/library';
 import { Account } from 'network/shapes/Account';
-import { Contribution, Goal } from 'network/shapes/Goal';
+import { Contribution, Goal } from 'network/shapes/Goals';
 import { DetailedEntity } from 'network/shapes/utils';
 import { parseQuantity } from 'network/shapes/utils/parse';
 import { ActionBar } from './ActionBar';
