+++ packages/client/src/network/shapes/Goals/index.ts
@@ -1,0 +1,5 @@
+export { canClaim, canContribute } from './functions';
+export { getAllGoals, getContributionByHash, getContributions, getGoalByIndex } from './getters';
+export { getGoal } from './types';
+
+export type { Contribution, Goal, Tier } from './types';
